+++ src/Vortice.Direct3D12/ID3D12GraphicsCommandList4.cs
@@ -23,9 +23,9 @@ public unsafe partial class ID3D12GraphicsCommandList4
     /// <param name="renderTargets">An array of <see cref="RenderPassRenderTargetDescription"/>, which describes bindings (fixed for the duration of the render pass) to one or more render target views (RTVs), as well as their beginning and ending access characteristics.</param>
     /// <param name="depthStencil">An optional <see cref="RenderPassDepthStencilDescription"/>, which describes a binding (fixed for the duration of the render pass) to a depth stencil view (DSV), as well as its beginning and ending access characteristics.</param>
     /// <param name="flags">The nature/requirements of the render pass; for example, whether it is a suspending or a resuming render pass, or whether it wants to write to unordered access view(s).</param>
-    public void BeginRenderPass(int renderTargetCount, RenderPassRenderTargetDescription[] renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
+    public void BeginRenderPass(uint renderTargetCount, RenderPassRenderTargetDescription[] renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
     {
-        Span<RenderPassRenderTargetDescription.__Native> renderTargetsNative = stackalloc RenderPassRenderTargetDescription.__Native[renderTargetCount];
+        Span<RenderPassRenderTargetDescription.__Native> renderTargetsNative = stackalloc RenderPassRenderTargetDescription.__Native[(int)renderTargetCount];
         for (int i = 0; i < renderTargetCount; ++i)
         {
             renderTargets[i].__MarshalTo(ref renderTargetsNative[i]);
@@ -51,7 +51,7 @@ public unsafe partial class ID3D12GraphicsCommandList4
     /// <param name="flags">The nature/requirements of the render pass; for example, whether it is a suspending or a resuming render pass, or whether it wants to write to unordered access view(s).</param>
     public void BeginRenderPass(RenderPassRenderTargetDescription[] renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
     {
-        BeginRenderPass(renderTargets.Length, renderTargets, depthStencil, flags);
+        BeginRenderPass((uint)renderTargets.Length, renderTargets, depthStencil, flags);
     }
 
     /// <summary>
@@ -62,9 +62,9 @@ public unsafe partial class ID3D12GraphicsCommandList4
     /// <param name="renderTargets">An array of <see cref="RenderPassRenderTargetDescription"/>, which describes bindings (fixed for the duration of the render pass) to one or more render target views (RTVs), as well as their beginning and ending access characteristics.</param>
     /// <param name="depthStencil">An optional <see cref="RenderPassDepthStencilDescription"/>, which describes a binding (fixed for the duration of the render pass) to a depth stencil view (DSV), as well as its beginning and ending access characteristics.</param>
     /// <param name="flags">The nature/requirements of the render pass; for example, whether it is a suspending or a resuming render pass, or whether it wants to write to unordered access view(s).</param>
-    public void BeginRenderPass(int renderTargetCount, Span<RenderPassRenderTargetDescription> renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
+    public void BeginRenderPass(uint renderTargetCount, Span<RenderPassRenderTargetDescription> renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
     {
-        Span<RenderPassRenderTargetDescription.__Native> renderTargetsNative = stackalloc RenderPassRenderTargetDescription.__Native[renderTargetCount];
+        Span<RenderPassRenderTargetDescription.__Native> renderTargetsNative = stackalloc RenderPassRenderTargetDescription.__Native[(int)renderTargetCount];
         for (int i = 0; i < renderTargetCount; ++i)
         {
             renderTargets[i].__MarshalTo(ref renderTargetsNative[i]);
@@ -90,7 +90,7 @@ public unsafe partial class ID3D12GraphicsCommandList4
     /// <param name="flags">The nature/requirements of the render pass; for example, whether it is a suspending or a resuming render pass, or whether it wants to write to unordered access view(s).</param>
     public void BeginRenderPass(Span<RenderPassRenderTargetDescription> renderTargets, RenderPassDepthStencilDescription? depthStencil = null, RenderPassFlags flags = RenderPassFlags.None)
     {
-        BeginRenderPass(renderTargets.Length, renderTargets, depthStencil, flags);
+        BeginRenderPass((uint)renderTargets.Length, renderTargets, depthStencil, flags);
     }
 
     /// <summary>
@@ -109,7 +109,7 @@ public unsafe partial class ID3D12GraphicsCommandList4
     /// <param name="postbuildInfoDescriptions">Array of descriptions for post-build info to generate describing properties of the acceleration structure that was built.</param>
     public void BuildRaytracingAccelerationStructure(BuildRaytracingAccelerationStructureDescription description, RaytracingAccelerationStructurePostBuildInfoDescription[] postbuildInfoDescriptions)
     {
-        BuildRaytracingAccelerationStructure(ref description, postbuildInfoDescriptions.Length, postbuildInfoDescriptions);
+        BuildRaytracingAccelerationStructure(ref description, (uint)postbuildInfoDescriptions.Length, postbuildInfoDescriptions);
     }
 
     public void InitializeMetaCommand(ID3D12MetaCommand metaCommand)
@@ -136,6 +136,6 @@ public unsafe partial class ID3D12GraphicsCommandList4
         RaytracingAccelerationStructurePostBuildInfoDescription description,
         ulong[] sourceAccelerationStructureData)
     {
-        EmitRaytracingAccelerationStructurePostbuildInfo(description, sourceAccelerationStructureData.Length, sourceAccelerationStructureData);
+        EmitRaytracingAccelerationStructurePostbuildInfo(description, (uint)sourceAccelerationStructureData.Length, sourceAccelerationStructureData);
     }
 }
